Syscalls: Cleanup headers, add names, add NotImplemented
