+++ External/FEXCore/Source/CMakeLists.txt
@@ -161,6 +161,7 @@ set (SRCS
   Interface/HLE/Syscalls/Thread.cpp
   Interface/HLE/Syscalls/Time.cpp
   Interface/HLE/Syscalls/Timer.cpp
+  Interface/HLE/Syscalls/NotImplemented.cpp
   Interface/Memory/MemMapper.cpp
   Interface/Memory/SharedMem.cpp
   Interface/IR/IR.cpp
